refactor: move game to lib/games
