+++ apps/web/src/lib/games/about-me/base/Action.ts
@@ -1,8 +1,10 @@
 import type { Game } from "../Game";
 
 export abstract class Action {
+	/** The unique action id */
 	public readonly id: string;
 
+	/** The action name visible to the player */
 	public readonly name: string;
 
 	public constructor(id: string, name: string) {
@@ -17,6 +19,10 @@ export abstract class Action {
 		};
 	}
 
+	/**
+	 * Called when the user is trying to perform this action
+	 * @param game The game instance
+	 */
 	public abstract handle(game: Game): GameReturnType;
 }
