+++ apps/web/src/lib/games/about-me/base/Result.ts
@@ -1,4 +1,5 @@
 export abstract class Result {
+	/** The unique result id */
 	public readonly id: string;
 
 	public constructor(id: string) {
