+++ apps/web/src/lib/games/about-me/base/Room.ts
@@ -7,6 +7,7 @@ export abstract class Room {
 	/** The name of the room */
 	public readonly name: string;
 
+	/** The game instance */
 	public readonly game: Game;
 
 	public constructor(name: string, game: Game) {
@@ -14,9 +15,33 @@ export abstract class Room {
 		this.game = game;
 	}
 
+	/**
+	 * The image of the current room
+	 * @param action The action the user is trying to perform
+	 */
 	public abstract image(action?: Action): string;
+
+	/**
+	 * The action the user is trying to perform
+	 * @param action The action the user is trying to perform
+	 */
 	public abstract examine(action?: Action): TextResult | undefined;
+
+	/**
+	 * The actions the player is able to use
+	 * @param action The action the user is trying to perform
+	 */
 	public abstract actions(action?: Action): Action[] | undefined;
+
+	/**
+	 * The action objects (options) the player is able to use
+	 * @param action The action the user is trying to perform
+	 */
 	public abstract objects(action?: Action): Action[] | undefined;
+
+	/**
+	 * Handles the custom action the user is trying to perform
+	 * @param action The action the user is trying to perform
+	 */
 	public abstract custom(action?: CustomAction): GameReturnType;
 }
